refactor: Split blueprint acquisition into source and cache modules
- blueprint/source-npm.ts: fetch @seamapi/types from the npm registry,
  extract the OpenAPI module, and build a blueprint from it.
- blueprint/cache.ts: the on-disk blueprint cache with TTL and version
  invalidation, atomic writes, and blueprint-version discovery.
- blueprint/source-remote.ts: build a blueprint from the OpenAPI
  document served by the configured server.
- blueprint/index.ts: getApiBlueprint source selector (was
  get-api-blueprint.ts).
- blueprint/endpoint.ts: command path -> endpoint lookup and response
  key (was get-command-blueprint-def.ts + get-response-key.ts).
