+++ src/lib/blueprint/endpoint.ts
@@ -1,5 +1,19 @@
-import { getCommandBlueprintDef } from './get-command-blueprint-def.js'
-import type { ContextHelpers } from './types.js'
+import type { ContextHelpers } from '../types.js'
+
+export const getCommandBlueprintDef = (
+  cmd: string[],
+  helpers: ContextHelpers,
+) => {
+  const path = `/${cmd.join('/').replace(/-/g, '_')}`
+  const def = helpers.blueprint.routes
+    .flatMap((route) => route.endpoints)
+    .find((endpoint) => endpoint.path === path)
+  if (!def) {
+    throw new Error(`No definition for path ${path}`)
+  }
+
+  return def
+}
 
 /**
  * The top level response key documented for a command,
