+++ src/lib/types.ts
@@ -1,5 +1,5 @@
 import type { Interactivity } from './args/parse.js'
-import type { ApiBlueprint } from './get-api-blueprint.js'
+import type { ApiBlueprint } from './blueprint/index.js'
 
 export interface ContextHelpers {
   blueprint: ApiBlueprint
